salvando para piter

+++ Tcc/Entrada/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 # Create your models here.
 class Mensage(models.Model):
-    id = models.AutoField(primary_key=True)
+    id = models.DecimalField(primary_key=True, max_digits=20, decimal_places=0)
     mensage = models.CharField(max_length=400, blank=True)
     id_user = models.CharField(max_length=150, blank=True)
     key = models.ForeignKey("self", related_name="self", on_delete=models.CASCADE, null=True, blank=True)
